Support optional CASA columns
